+++ baka/__init__.py
@@ -8,7 +8,7 @@
     :license: BSD, see LICENSE for more details.
 """
 
-__version__ = '0.4.3'
+__version__ = '0.4.3.dev1'
 
 from .app import Baka
 from .log import log